requirements.txt
streamlit
openai>=1.66.0
python-dotenv
plotly
pandas
